moving the MySQL collation and charset tests
